+++ sdk/resourcemanager/azure-resourcemanager-network/src/main/java/com/azure/resourcemanager/network/fluent/models/PrivateEndpointConnectionInner.java
@@ -153,6 +153,13 @@ public class PrivateEndpointConnectionInner extends SubResource {
         return this.linkIdentifier;
     }
 
+    /** {@inheritDoc} */
+    @Override
+    public PrivateEndpointConnectionInner withId(String id) {
+        super.withId(id);
+        return this;
+    }
+
     /**
      * Validates the instance.
      *