+++ sdk/resourcemanager/azure-resourcemanager-network/src/main/java/com/azure/resourcemanager/network/fluent/models/VirtualNetworkGatewayInner.java
@@ -11,6 +11,7 @@ import com.azure.core.management.SubResource;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.network.models.AddressSpace;
 import com.azure.resourcemanager.network.models.BgpSettings;
+import com.azure.resourcemanager.network.models.ExtendedLocation;
 import com.azure.resourcemanager.network.models.ProvisioningState;
 import com.azure.resourcemanager.network.models.VirtualNetworkGatewaySku;
 import com.azure.resourcemanager.network.models.VirtualNetworkGatewayType;
@@ -20,6 +21,7 @@ import com.azure.resourcemanager.network.models.VpnType;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
+import java.util.Map;
 
 /** A common class for general resource information. */
 @JsonFlatten
@@ -137,6 +139,19 @@ public class VirtualNetworkGatewayInner extends Resource {
      */
     @JsonProperty(value = "properties.inboundDnsForwardingEndpoint", access = JsonProperty.Access.WRITE_ONLY)
     private String inboundDnsForwardingEndpoint;
+
+    /*
+     * MAS FIJI customer vnet resource id. VirtualNetworkGateway of type local
+     * gateway is associated with the customer vnet.
+     */
+    @JsonProperty(value = "properties.virtualNetworkExtendedLocationResourceId")
+    private String virtualNetworkExtendedLocationResourceId;
+
+    /*
+     * The extended location of type local virtual network gateway.
+     */
+    @JsonProperty(value = "properties.extendedLocation")
+    private ExtendedLocation extendedLocation;
 
     /*
      * Resource ID.
@@ -455,6 +470,49 @@ public class VirtualNetworkGatewayInner extends Resource {
     }
 
     /**
+     * Get the virtualNetworkExtendedLocationResourceId property: MAS FIJI customer vnet resource id.
+     * VirtualNetworkGateway of type local gateway is associated with the customer vnet.
+     *
+     * @return the virtualNetworkExtendedLocationResourceId value.
+     */
+    public String virtualNetworkExtendedLocationResourceId() {
+        return this.virtualNetworkExtendedLocationResourceId;
+    }
+
+    /**
+     * Set the virtualNetworkExtendedLocationResourceId property: MAS FIJI customer vnet resource id.
+     * VirtualNetworkGateway of type local gateway is associated with the customer vnet.
+     *
+     * @param virtualNetworkExtendedLocationResourceId the virtualNetworkExtendedLocationResourceId value to set.
+     * @return the VirtualNetworkGatewayInner object itself.
+     */
+    public VirtualNetworkGatewayInner withVirtualNetworkExtendedLocationResourceId(
+        String virtualNetworkExtendedLocationResourceId) {
+        this.virtualNetworkExtendedLocationResourceId = virtualNetworkExtendedLocationResourceId;
+        return this;
+    }
+
+    /**
+     * Get the extendedLocation property: The extended location of type local virtual network gateway.
+     *
+     * @return the extendedLocation value.
+     */
+    public ExtendedLocation extendedLocation() {
+        return this.extendedLocation;
+    }
+
+    /**
+     * Set the extendedLocation property: The extended location of type local virtual network gateway.
+     *
+     * @param extendedLocation the extendedLocation value to set.
+     * @return the VirtualNetworkGatewayInner object itself.
+     */
+    public VirtualNetworkGatewayInner withExtendedLocation(ExtendedLocation extendedLocation) {
+        this.extendedLocation = extendedLocation;
+        return this;
+    }
+
+    /**
      * Get the id property: Resource ID.
      *
      * @return the id value.
@@ -471,6 +529,20 @@ public class VirtualNetworkGatewayInner extends Resource {
      */
     public VirtualNetworkGatewayInner withId(String id) {
         this.id = id;
+        return this;
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public VirtualNetworkGatewayInner withLocation(String location) {
+        super.withLocation(location);
+        return this;
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public VirtualNetworkGatewayInner withTags(Map<String, String> tags) {
+        super.withTags(tags);
         return this;
     }
 
@@ -494,6 +566,9 @@ public class VirtualNetworkGatewayInner extends Resource {
         }
         if (customRoutes() != null) {
             customRoutes().validate();
+        }
+        if (extendedLocation() != null) {
+            extendedLocation().validate();
         }
     }
 }