+++ sdk/spring/azure-spring-cloud-starter-cache/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Release History
 
-## 2.0.0-beta.2 (Unreleased)
+## 2.2.0-beta.1 (Unreleased)
 
+
+## 2.0.0 (2020-12-30)
+### Breaking Changes
+- Deprecated the `spring.cloud.azure.managed-identity.client-id` property,
+  use `spring.cloud.azure.client-id` to set the managed identity id when using Managed Identity.
 
 ## 2.0.0-beta.1 (2020-11-18)
 ### Breaking Changes